pyproject.toml
[tool.poetry]
name = "ssmerge46"
version = "0.1.0"
description = "A discod bot that merges screenshots taken in 'Umamusume Pretty Derby'"
authors = ["susu8989"]
readme = "README.md"
packages = [{include = "ssmerge46", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9.0"

"discord.py" = "^2.2.3"
opencv-python = "^4.7.0"
numpy = "^1.24.3"
mashumaro = "^3.7"
pyyaml = "*"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.2.0"
pydocstring = "^0.2.1"
pydocstyle= "^6.3.0"
pyproject-flake8 = "^6.0.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
